Trying to use data Table
I was trying to use same method that I mostly use in Specflow  feature files with a lot of data that i want to insert or to compare.
Sad thing was that, in the tutorial i watched, there was used a method called table.row which doesn't exist anymore. After few seconds I found that actuall method with the same incrementations is table.cells in Cucumber.